requirements.txt
Django>=1.2.3
django_annoying>=0.7.4
-e git+git://github.com/odesk/python-odesk.git#egg=odesk-0.1.1
-e git+git://github.com/solex/django-odesk.git#egg=django_odesk-0.1.1
